Better folder structure and build script for the future.
