+++ run.bat
@@ -1,0 +1,1 @@
+.\bin\bob16.exe
